Fix test linker errors: extern scope in Boost.Test namespace
Move extern declarations to file scope so they resolve to global
symbols instead of the Boost test suite namespace. Cast static const
member to avoid ODR address requirement.

+++ src/test/serialization_tests.cpp
@@ -146,7 +146,7 @@ BOOST_AUTO_TEST_CASE(block_header_roundtrip)
 
 BOOST_AUTO_TEST_CASE(block_current_version)
 {
-    BOOST_CHECK_EQUAL(CBlock::CURRENT_VERSION, 6);
+    BOOST_CHECK_EQUAL((int)CBlock::CURRENT_VERSION, 6);
 }
 
 // --- CDiskTxPos ---
